refactor(combat-screen): extract detail sub-components, lucide clear icon, header tests
- Replace bare × glyph in temp-HP Clear button with <X size={11} aria-hidden /> (lucide)
- Extract Concentration+Reaction chips from DetailVitals into ConcentrationReaction.tsx
- Extract DM disposition/visibility/alias controls from DetailHeader into HeaderControls.tsx
- Add tests: player max-HP static text, Enemy disposition onUpdate, Remove confirm/cancel flows
